server/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
aries-askar==0.3.2
base58==2.1.1
black==24.8.0
cached-property==1.5.2
canonicaljson==2.0.0
cffi==1.17.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.0
h11==0.14.0
idna==3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyNaCl==1.5.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.5
